directory 구조변경 및 refactoring
