+++ app/lib/api/auth/authKakao.dart
@@ -2,10 +2,9 @@ import 'package:dio/dio.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
 class AuthKakao {
-  Future<Response> authKakao() async {
+  static Future<Response> authKakao() async {
     await dotenv.load(fileName: ".env");
     String? apiKey = dotenv.env['API_KEY'];
-
     final res = await Dio().get('$apiKey/auth/kakao');
     return res;
   }
