+++ app/lib/pages/profile/profile_modify.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:pet/common/component/appbars/go_back_appbar.dart';
-import 'package:pet/common/component/next_button.dart';
+import 'package:pet/common/component/buttons/next_button.dart';
 import 'package:pet/const/regions.dart';
 import 'package:pet/pages/profile/profile_main.dart';
-import '../../common/component/dropdown_button.dart';
+import '../../common/component/buttons/dropdown_button.dart';
 import '../../style/colors.dart';
 
 class ProfileModify extends StatelessWidget {
